Udpate final exercise Part 4

+++ controllers/blogs.js
@@ -1,5 +1,16 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blogs')
+const User = require('../models/user')
+
+// const jwt = require('jsonwebtoken')
+
+// const getTokenFrom = req => {
+//   const authorization = req.get('authorization')
+//   if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+//     return authorization.substring(7)
+//   }
+//   return null
+// }
 
 // blogsRouter.get('/', (request, response, next) => {
 //   Blog
@@ -11,7 +22,7 @@ const Blog = require('../models/blogs')
 // })
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user',{username: 1, name: 1})
   response.json(blogs)
 })
 
@@ -34,14 +45,28 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+
+  // const token = getTokenFrom(request)
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  // if (!request.token || !decodedToken.id) {
+  //   return response.status(401).json({error: 'token missing or invalid'})
+  // }
+  // // const user = await User.findById(body.userId)
+  // const user = await User.findById(decodedToken.id)
+  const user = request.user
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: user._id
   })
-  const result = await blog.save()
-  response.status(201).json(result)
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.status(201).json(savedBlog)
 })
 
 // blogsRouter.put('/:id',(req, res, next) => {
@@ -82,7 +107,32 @@ blogsRouter.put('/:id', async (req, res) => {
 // })
 
 blogsRouter.delete('/:id', async (req, res) => {
-  await Blog.findByIdAndRemove(req.params.id)
+
+  // const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  // if (!req.token || !decodedToken.id) {
+  //   return res.status(401).json({erorr: 'token misisng or invalid'})
+  // }
+  // // Pass authentication then check user name
+  // const user = await User.findById(decodedToken.id)
+
+  const user = req.user
+  // Get the user id who create blog
+  const blog = await Blog.findById(req.params.id)
+  // Check if the username has the right to delete the blog
+  if (!user || !blog) {
+    return res.status(400).json({error:'Invalid User or Resource does not exist'})
+  }
+  if (user._id.toString() === blog.user.toString()) {
+    // If matched then remove the blog
+    await Blog.findByIdAndRemove(req.params.id)
+    // Then remove the blog in the user schema
+    user.blogs = user.blogs.filter(blogs => blogs.id !== req.params.id)
+    await user.save()
+    res.status(204).end()
+  } else {
+    res.status(401).json({error:'Unauthorized Action'})
+  }
+  
   res.status(204).end()
 })
